% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/damage_calc.R
\name{damage_calc}
\alias{damage_calc}
\title{Calculate damage using the Gen 1 Pokémon formula (simplified)}
\usage{
damage_calc(attacker, defender, attacker_level, power)
}
\arguments{
\item{attacker}{The name of the attacking Pokémon}

\item{defender}{The name of the defending Pokémon}

\item{attacker_level}{The level of the attacking Pokémon (manual input)}

\item{power}{The base power of the move being used (manual input)}
}
\value{
The amount of damage dealt
}
\description{
This function calculates how much damage one Pokémon does to another
using a simplified version of the original Generation 1 formula
It pulls base stats from the 'pokemon' package based on attacker and defender names
It does not include STAB or type effectiveness
Critical hits are based on the attacker's speed
Critical hits double the attacker's level for the damage calculation
}
