% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_battle.R
\name{simulate_battle}
\alias{simulate_battle}
\title{Simulates a pokemon battle where the attacker uses the same move until the
defender is KO'd, option to simulate many at once by specifying 'n'}
\usage{
simulate_battle(
  attacker,
  defender,
  attacker_level,
  defender_level,
  power,
  n = 1,
  print = TRUE
)
}
\arguments{
\item{attacker}{The name of the attacking Pokémon}

\item{defender}{The name of the defending Pokémon}

\item{attacker_level}{The level of the attacking Pokémon (manual input)}

\item{defender_level}{The level of the defending Pokémon (manual input)}

\item{power}{The base power of the move (manual input)}

\item{n}{Number of simulations to run (default is 1)}

\item{print}{Whether to print the results automatically (default is TRUE).
Set to FALSE when running a large number of simulations to avoid console clutter}
}
\value{
A list containing turn counts damage logs and HP logs
}
\description{
This function uses damage_calc() to simulate how many turns it takes for
the attacker to KO the defender using a specific move
}
